requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
click==8.0.3
databases==0.5.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.2
pyasn1==0.4.8
pydantic==1.8.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.6.0
